% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Static Data.R
\name{add_spread}
\alias{add_spread}
\title{Add Spread}
\usage{
add_spread(input)
}
\arguments{
\item{input}{standardised post trade}
}
\value{
post trade with spread column added.
}
\description{
Add an indicative spread column to the standard post trade.
}
