% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic Analysis.R
\name{basic_analysis}
\alias{basic_analysis}
\title{Basic Analysis}
\usage{
basic_analysis(input, spread_normalise = FALSE, ...)
}
\arguments{
\item{input}{data.frame containing the standardised post trade}
}
\value{
analysis data.frame
}
\description{
Simple performance analysis
}
