% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic Analysis.R
\name{equal_freq_analysis}
\alias{equal_freq_analysis}
\title{Equal Freq Analysis}
\usage{
equal_freq_analysis(
  input,
  break_var,
  num_groups,
  ...,
  spread_normalise = FALSE
)
}
\arguments{
\item{input}{data.frame containing the standardised post trade}

\item{break_var}{variable to break on}

\item{num_groups}{number of groups to split into}

\item{spread_normalise}{return benchmarks normalised by the average spread}
}
\value{
analysis data.frame
}
\description{
Performance analysis with equal freq bins of a continuous variable
}
