% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Java Date Format.R
\name{extract_date}
\alias{extract_date}
\title{Convert java date / time into a Date}
\usage{
extract_date(java_date_strings)
}
\arguments{
\item{java_date_strings}{list of datetimes}
}
\value{
parsed Dates
}
\description{
Convert java date / time into a Date
}
