% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Java Date Format.R
\name{extract_date_time}
\alias{extract_date_time}
\title{Convert java date / time into a PosixCT}
\usage{
extract_date_time(java_date_strings, tz = "Australia/Sydney")
}
\arguments{
\item{java_date_strings}{list of datetimes}

\item{tz}{timezone to use}
}
\value{
parsed PosixCTs
}
\description{
Convert java date / time into a PosixCT
}
