% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Actual Vol Profile Loader.R
\name{load_actual_profile}
\alias{load_actual_profile}
\title{Create a profile of the actual traded volume}
\usage{
load_actual_profile(syms, dates)
}
\arguments{
\item{syms}{list of syms}

\item{dates}{list of dates}
}
\value{
volume profile per sym / date
}
\description{
This will add in a dummy point for open and close auctions
}
