% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loader.R
\name{load_multiple}
\alias{load_multiple}
\title{Load Multiple Post Trades}
\usage{
load_multiple(
  path,
  pattern = "*.xlsx",
  add_filename = FALSE,
  sheet_name = "Detail",
  row_offset = NA
)
}
\arguments{
\item{path}{path containing the files to read}

\item{pattern}{pattern for files to match}

\item{add_filename}{should the file name be added to the returned data.frame}

\item{sheet_name}{sheet name to load (excel only)}

\item{row_offset}{offset to use when loading data}
}
\value{
data.frame containing the loaded files.
}
\description{
Load multiple post trade files (csv and excel supported) returning a single
data.frame.
}
