% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolProfileLoader.R
\name{load_vol_profile}
\alias{load_vol_profile}
\title{Load a volume profile using cached data}
\usage{
load_vol_profile(syms, dates)
}
\arguments{
\item{syms}{list of syms}

\item{dates}{list of dates}
}
\value{
volume profile per sym / date
}
\description{
This will adjust the batch volume profile, adding in a dummy point for open and close auctions
}
