% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pairwise.R
\name{pairwise_analysis}
\alias{pairwise_analysis}
\title{Pairwise Analysis}
\usage{
pairwise_analysis(
  input,
  benchmark_name = "arrival_dev",
  group_variables = c("date", "sym", "side", "order_size"),
  complete_cases = TRUE,
  batch_id = "batch",
  add_diff = FALSE
)
}
\arguments{
\item{input}{data.frame containing the standardised post trade}

\item{benchmark_name}{name of the benchmark to analyse}

\item{complete_cases}{should we only return data where results are present for all batches.}

\item{add_diff}{add a difference column (only for cases where there are two batches)}

\item{group_variable}{list of varibles to uniquely identify a single run}

\item{batch}{column containing an indicator of the run}
}
\value{
analysis data.frame
}
\description{
Analysis of multiple runs.
}
