% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceDecomposition.R
\name{performance_decomposition}
\alias{performance_decomposition}
\title{Run profile decomposition for simulations.}
\usage{
performance_decomposition(file_name, return_full = FALSE)
}
\arguments{
\item{file_name}{path to the bin_data export}

\item{return_full}{return the raw profile bins}
}
\value{
decomposition of the slippage into historical, tracking and execution components.
}
\description{
Run profile decomposition for simulations.
}
