% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardise.R
\name{standardise_post_trade}
\alias{standardise_post_trade}
\title{Standardise post trade}
\usage{
standardise_post_trade(input)
}
\arguments{
\item{input}{data.frame containing post trade to be standardised}
}
\value{
standardised data.frame
}
\description{
Attempt standardisation of column names and formats.
}
